requirements.txt
langchain-community>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.2
PyPDF2>=3.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
python-dotenv>=1.0.0
tiktoken>=0.5.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2 